packages/api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
biopython>=1.83
numpy>=1.26.0
httpx>=0.27.0
